move Mutex to base signer; improve log print and unit test
