+++ testutil/sample/os.go
@@ -1,4 +1,4 @@
-package testutils
+package sample
 
 import (
 	"os"
