+++ zetaclient/chains/base/logger_test.go
@@ -5,9 +5,9 @@ import (
 
 	"github.com/stretchr/testify/require"
 
+	"github.com/zeta-chain/zetacore/testutil/sample"
 	"github.com/zeta-chain/zetacore/zetaclient/chains/base"
 	"github.com/zeta-chain/zetacore/zetaclient/config"
-	"github.com/zeta-chain/zetacore/zetaclient/testutils"
 )
 
 func TestInitLogger(t *testing.T) {
@@ -24,7 +24,7 @@ func TestInitLogger(t *testing.T) {
 				LogFormat: "json",
 				LogLevel:  1, // zerolog.InfoLevel,
 				ComplianceConfig: config.ComplianceConfig{
-					LogPath: testutils.CreateTempDir(t),
+					LogPath: sample.CreateTempDir(t),
 				},
 			},
 			fail: false,
@@ -35,7 +35,7 @@ func TestInitLogger(t *testing.T) {
 				LogFormat: "text",
 				LogLevel:  2, // zerolog.WarnLevel,
 				ComplianceConfig: config.ComplianceConfig{
-					LogPath: testutils.CreateTempDir(t),
+					LogPath: sample.CreateTempDir(t),
 				},
 			},
 			fail: false,
@@ -46,7 +46,7 @@ func TestInitLogger(t *testing.T) {
 				LogFormat: "unknown",
 				LogLevel:  3, // zerolog.ErrorLevel,
 				ComplianceConfig: config.ComplianceConfig{
-					LogPath: testutils.CreateTempDir(t),
+					LogPath: sample.CreateTempDir(t),
 				},
 			},
 			fail: false,
@@ -58,7 +58,7 @@ func TestInitLogger(t *testing.T) {
 				LogLevel:   4, // zerolog.DebugLevel,
 				LogSampler: true,
 				ComplianceConfig: config.ComplianceConfig{
-					LogPath: testutils.CreateTempDir(t),
+					LogPath: sample.CreateTempDir(t),
 				},
 			},
 		},
